Inventory/requirements.txt
arabic-reshaper
asgiref
asn1crypto
certifi
cffi
chardet
charset-normalizer
click
colorama
cryptography
cssselect2
Django
djangorestframework
gunicorn
html5lib
idna
lxml
oscrypto
packaging
pillow
pycparser
pyHanko
pyhanko-certvalidator
pypdf
pypng
python-bidi
python-dotenv
pytz
PyYAML
qrcode
reportlab
requests
six
sqlparse
svglib
tinycss2
typing_extensions
tzdata
tzlocal
uritools
urllib3
webencodings
xhtml2pdf
whitenoise